+++ packages/slots/.storybook/preview.ts
@@ -1,8 +1,8 @@
 import './storybook-base.css';
 import '@plone/components/dist/basic.css';
-import '../main.css';
+import '../styles/main.css';
 import config from '@plone/registry';
-import installSlots from '../config';
+import installSlots from '../index';
 import installBlocks from '@plone/blocks';
 
 config.set('slots', {});