+++ packages/volto/cypress/tests/core/basic/a11y.js
@@ -5,6 +5,16 @@ describe('Accessibility Tests', () => {
   });
 
   it('Front page has not a11y violations', () => {
+    cy.configureAxe({
+      rules: [
+        // Disabled until we fix robotframework to install the volto distribution
+        // with example content
+        {
+          id: 'page-has-heading-one',
+          enabled: false,
+        },
+      ],
+    });
     cy.checkA11y(); // fail for a11y violations
   });