+++ packages/slots/CHANGELOG.md
@@ -8,4 +8,15 @@
 
 <!-- towncrier release notes start -->
 
+## 1.0.0-alpha.0 (2025-01-24)
+
+### Bugfix
+
+- Unify the ESlint config. @sneridagh [#6599](https://github.com/plone/volto/issues/6599)
+
+### Internal
+
+- Initial commit, first steps. @sneridagh [#6409](https://github.com/plone/volto/issues/6409)
+- Centralize `tsconfig`. @sneridagh [#6536](https://github.com/plone/volto/issues/6536)
+
 ## 1.0.0 (unreleased)